requirements.txt
numpy==1.24.4
opencv-python==4.7.0.72
nptyping==2.5.0
omegaconf==2.3.0
openvino==2023.0.0
hydra-core==1.3.2
